Edited historical rates screen
