+++ lib/pages/main_page.dart
@@ -1,6 +1,6 @@
 import 'package:currency_exchange_app/pages/landing_page.dart';
 import 'package:currency_exchange_app/pages/profile_page.dart';
-import 'package:currency_exchange_app/pages/test_chart.dart';
+import 'package:currency_exchange_app/pages/historical_rates_chart_page.dart';
 import 'package:currency_exchange_app/utils/colors.dart';
 import 'package:currency_exchange_app/utils/dimens.dart';
 import 'package:flutter/cupertino.dart';
@@ -18,7 +18,7 @@ class _MainPageState extends State<MainPage> {
 
   List<Widget> screenWidgets = [
     const LandingPage(),
-    const TestChart(),
+    const HistoricalRatesChart(),
     const ProfilePage()
   ];
   @override
